+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Exit codes according to documentation (https://nowarp.io/tools/misti/docs/tutorial/cli#exit-codes): Issue [#12](https://github.com/nowarp/blueprint-misti/issues/12)
+
 ## [0.6.0] - 2024-12-22
 
 ### Added